// src/types/RouletteTypes.ts
export enum RouletteState {
  RESET = 'reset',
  RESETTING = 'resetting',
  IDLE = 'idle',
  WAITING = 'waiting',
  LOAD = 'load',
  LOADING = 'loading',
  READY = 'ready',
  SPIN = 'spin',
  SPINING = 'spining',
  SHOOTING = 'shooting',
  SHOT = 'shot'
}

export enum LoadState {
  EMPTY = 'empty',
  HALF = 'half',
  FULL = 'full'
}